+++ src/react/App/pages/HomePage/HomePage.tsx
@@ -1,19 +1,28 @@
 import React from 'react'
 import styles from './styles'
-import WelcomeRegisterBox from '../../../components/modules/WelcomeRegisterBox/WelcomeRegisterBox'
+import WelcomeRegisterBox, { RegistrationData } from '../../../components/modules/WelcomeRegisterBox/WelcomeRegisterBox'
+import { loginApiMutation, signUpApiMutation } from '../../../apis/signon'
 
 const HomePage = () => {
   const c = styles()
+  const signUpApi = signUpApiMutation.make()
+  const loginApi = loginApiMutation.make()
 
-  function handleOnRegister (): boolean {
-    return true
+  async function registerNewUser ({ email, password }: RegistrationData) {
+    await signUpApi.call({
+      data: { email, password }
+    }, {
+      onSuccess: data => null,
+      onError: error => null
+    })
   }
 
   return <div className={c.root}>
-    <WelcomeRegisterBox onRegisterSuccess={handleOnRegister}/>
+    <WelcomeRegisterBox onValidRegistration={registerNewUser}/>
 
   </div>
 }
+
 
 export default HomePage