+++ src/react/apis/signon.ts
@@ -1,0 +1,5 @@
+import { loginController, signUpController } from '../constants/urls'
+import ApiMutationFactory from './ApiMutation'
+
+export const signUpApiMutation = new ApiMutationFactory({ url: signUpController, method: 'POST' })
+export const loginApiMutation = new ApiMutationFactory({ url: loginController, method: 'POST' })