+++ src/react/util/log.tsx
@@ -1,0 +1,6 @@
+const log = (message?: unknown, ...optionalParams: unknown[]): void => {
+  console.log(message, optionalParams)
+}
+
+export default log
+